Drop Python 3.5 from CI
I forgot about the limitations imposed by supporting Python 3.5 and added
a lot of fixtures in the test suite that use pathlib. Rather than rip all
of that work our of the test suite, I am droping CI for Python 3.5. This is
justified because the only place we use the package under Python 3.5 is on
orincus for the nowcast-agirf runs, and I have to live with the consequences
of maintaining less well support code there.