Change memory per node for `sbatch` runs to 0
This is to ensure that jobs go on to the "by-node" queue on `graham`.
The queue selection algorithm in the scheduler on `graham` appears to have
changed (without announcement) on 13-Jul-2021 so that the previous memory per
node of 125G now sends jobs to the "by-core" queue which is generally more
crowded.