Finalize example HPC module versions re: StdEnv/2020
Compute Canada changed the standard software environment to StdEnv/2020
on 1-Apr-2021. It took several months for things to stabilize to at state
in which NEMO and XIOS-2 could run in the new environment.

NEMO-Cmd doesn't care directly about the modules; that is the user's
responsibility to specify in the run description YAML file.
This revision updates the docs and unit tests to show the new recommended
versions of modules to load for runs on the Compute Canada clusters.