Change project badges layout in dev docs & README (#72)
* Change project badges layout in dev docs & README

Reformatted the project badges in `README.rst` and `pkg_development.rst` into a
table to improve readability and clarity. Badges are now grouped by
"Continuous Integration", "Documentation", "Package" and "Meta" to provide a
better overview of the project status, and its development tools and processes.

* Split Meta cell in badges table

This is meant to improve readability, especially in Sphinx HTML rendering of
docs.