Update to Python 3.13 (#92)
* Add Python 3.13.0rc2 to GHA pytest workflow

Added Python 3.13.0rc2 to the pytest-with-coverage workflow python-version
matrix to test compatibility with the latest release candidate. Also added
'conda-forge/label/python_rc' channel to the test environment configuration
to enable installation of the release candidate.

* Change to Python 3.13 in GHA pytest workflow

Changed from Python 3.13.0rc2 to 3.13 in the pytest-with-coverage workflow
python-version matrix to test compatibility with the final release. Also dropped
the 'conda-forge/label/python_rc' channel from the test environment configuration
because it is no longer needed.

* Update ReadTheDocs build config

Upgraded the OS to Ubuntu 24.04 and updated the Python toolchain to
mambaforge-23.11 in the .readthedocs.yaml file. This ensures compatibility with
the latest environment settings.

* Update to Python 3.13 in envs & GHA sphinx-linkcheck

Upgraded Python from version 3.12 to 3.13 across various YAML files. This
includes updates to development, RTD, and HPC environments, as well as GHA
sphinx-linkcheck workflow.

* Update Sphinx & optional dependencies version pins

Updated Sphinx and optional dependencies packages pinned versions in
`environment-dev.yaml`, `environment-test.yaml`, and `environment-rtd.yaml` to
get functional Python 3.13 environment.

* Update to Python 3.13 for development and use

Upgraded the required Python version to 3.13 in `pyproject.toml` and updated
related documentation and links to reflect this change.

* Revert Python version requirement to >=3.11

Reverted the required Python version from 3.13 to 3.11 in pyproject.toml
because I made a mistake. It has to be the oldest version that the package
supports, and we're leaving that at 3.11 due to several group members' envs
on graham.